Clarify eventlog and feed iterator options, and usage in tests
The `gt`/`gte`/`lt`/`lte` options for enventlog and feed iterators
take the hash of an item, and that wasn't clear from the API docs.
Also added more info to the `limit` option.

To see usage examples, many people are likely to turn to test suite.
In the tests for eventlog and feed, the `items` variable first
referred to an iterator with the actual items, and then later in
the tests `items` was an array of the item hashes. This switch of
types can be pretty confusing, thus the variable names were disambiguated.