docs: Add TOCs using an npm script
This is a stab at making #426 work automatically, so that we don't need to manually adjust the files. I've also edited some of the depths of the headings. What do you think, @mistakia?

One thing I would like is to add this to the build. It shouldn't change anything if there are no changes to make.